+++ backend/src/main/java/com/coronakarma/backend/CoronaKarmaBackendApplication.java
@@ -1,7 +1,5 @@
 package com.coronakarma.backend;
 
-import java.util.Arrays;
-
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
