+++ backend/src/main/java/com/coronakarma/backend/supporter/Supporter.java
@@ -1,12 +1,8 @@
-package com.coronakarma.backend.model;
+package com.coronakarma.backend.supporter;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
+import com.coronakarma.backend.common.Person;
+
+import javax.persistence.*;
 
 @Entity
 @Table(name = "supporter")
