+++ backend/src/main/java/com/coronakarma/backend/supporter/SupporterRepository.java
@@ -1,11 +1,9 @@
-package com.coronakarma.backend.repository;
+package com.coronakarma.backend.supporter;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-
-import com.coronakarma.backend.model.Supporter;
 
 public interface SupporterRepository extends JpaRepository<Supporter, Long> {
